Starting to fix type issues
